Thymeleaf and almost CRUD; to be cleaned up

+++ src/main/java/com/example/demo/repository/PersonRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.*;
 import com.example.demo.model.Person;
 
 @Repository
-public interface PersonRepository extends CrudRepository<Person, Long> {}
+public interface PersonRepository extends CrudRepository<Person, Integer> {}
